HDDS-16101. Extract bucket lifecycle operations into dedicated handler
